Added a dedicated settings page.
