+++ src/TF2HUD.Editor/Models/GitJson.cs
@@ -1,0 +1,12 @@
+﻿using System.Text.Json.Serialization;
+
+namespace HUDEditor.Models
+{
+    public class GitJson
+    {
+        [JsonPropertyName("name")] public string Name { get; set; }
+        [JsonPropertyName("sha")] public string SHA { get; set; }
+        [JsonPropertyName("download_url")] public string Download { get; set; }
+        [JsonPropertyName("type")] public string Type { get; set; }
+    }
+}
