+++ src/TF2HUD.Editor/Models/HudJson.cs
@@ -79,12 +79,4 @@ namespace HUDEditor.Models
         [JsonPropertyName("SpecialParameters")] public string[] SpecialParameters;
         [JsonPropertyName("Value")] public string Value;
     }
-
-    public class GitJson
-    {
-        [JsonPropertyName("name")] public string Name { get; set; }
-        [JsonPropertyName("sha")] public string SHA { get; set; }
-        [JsonPropertyName("download_url")] public string Download { get; set; }
-        [JsonPropertyName("type")] public string Type { get; set; }
-    }
 }
