+++ src/TF2HUD.Editor/SettingsWindow.xaml
@@ -1,0 +1,46 @@
+﻿<Window x:Class="HUDEditor.SettingsWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:lex="http://wpflocalizeextension.codeplex.com"
+        xmlns:classes="clr-namespace:HUDEditor.Classes"
+        lex:LocalizeDictionary.DesignCulture="en"
+        lex:ResxLocalizationProvider.DefaultAssembly="TF2HUD.Editor"
+        lex:ResxLocalizationProvider.DefaultDictionary="Resources"
+        mc:Ignorable="d"
+        Title="{lex:Loc ui_title}"
+        Width="1320"
+        Height="850"
+        WindowStartupLocation="CenterScreen"
+        FontFamily="../Resources/TF2Secondary.ttf #TF2 Secondary"
+        FontSize="20px"
+        Background="#2B2724">
+    <Window.Resources>
+        <classes:NullCheckConverter x:Key="NullCheckConverter" />
+        <classes:NotNullCheckConverter x:Key="NotNullCheckConverter" />
+        <classes:NullCheckConverterVisibility x:Key="NullCheckConverterVisibility" />
+        <classes:NotNullCheckConverterVisibility x:Key="NotNullCheckConverterVisibility" />
+        <classes:BtnInstallContentConverter x:Key="BtnInstallContentConverter" />
+        <classes:PageBackgroundConverter x:Key="PageBackgroundConverter" />
+    </Window.Resources>
+    <Grid>
+        <!-- App Info and Changelog -->
+        <ScrollViewer VerticalScrollBarVisibility="Hidden" Grid.Column="1" Grid.Row="0" Visibility="{Binding Json.HighlightedHud, Converter={StaticResource NotNullCheckConverterVisibility}}">
+            <StackPanel Margin="20">
+                <Image Source="Resources/banner.png" Height="142" Width="621" Margin="0,0,0,15" />
+                <TextBlock Style="{StaticResource BodyText}" FontSize="21" TextAlignment="Center">
+                            <Run Foreground="#70B04A" Text="Install" /><Run Text=" and" />
+                            <Run Foreground="#38F3AB" Text="customize" />
+                            <Run Text="your favorite Team Fortress 2 HUDs! For additional information and support, check out the" />
+                            <Run Foreground="#FFD700" Text="documentation" /><Run Text=" linked below." />
+                </TextBlock>
+                <Border Background="#A49E9E" HorizontalAlignment="Stretch" Height="1" Margin="0,15,0,10" />
+                <Label Content="Version 2.6" Style="{StaticResource HeaderText}" Margin="0,0,0,10" />
+                <TextBlock Style="{StaticResource BodyText}" FontSize="22" TextAlignment="Left">
+                    <Run Text="* TOOOOOODOOOOOOOO" /><LineBreak />
+                </TextBlock>
+            </StackPanel>
+        </ScrollViewer>
+    </Grid>
+</Window>
